Fix type confusion due to translation issues in the first chapter
This commit fix a "type confusion" due to translation issue in the first chapter
inside `Valores e tipos' section.
In Python, a comma is not used to separate thousands like it is used in english.
The example shows that something like `a = 1,000,000` creates valid python code,
but not with the same meaning as in engish.
The example creates a tuple with three integer elements, e.g.:
```
>>> a = 1,000,000
>>> a
(1, 0, 0)
>>> type(a)
<class 'tuple'>
```

This commit removes part of translation note that could generate confusion among
readers, it also changes the number `1.000.000' (which is a float in python) back to
`1,000,000' so it can be a tuple.